Embedded Product Approach
We have been able to achieve embedding the product on a html page which has the custom length ordering table. The price is being fetched from the embedded product. The add to cart button is adding qty = length of v slot in meters to the cart. Price being passed is equal to the price of the added lengths.